Rename class and restructure
